fix(fuzz): keep fuzz_exports round_trip helpers in production fse module
`fse::round_trip` and its `check_tables` helper are gated
`#[cfg(any(test, feature = "fuzz_exports"))]` and consumed by the `fse`
fuzz target through the crate's public surface, not test cases. The test
sweep moved them into `fse/tests.rs` (compiled only under `test`), so the
`fuzz_exports` build could no longer resolve `round_trip` and the
encoder's `into_table` / `encode` went dead. Move both back to
`fse/mod.rs`.
